Re-introduce a yplus-cutoff
The (small) eddies generated below this wall-distance are of too high
intensity (recall that volume integral over eddies is normalized).
Hence, they would require a supersmall timestep to avoid CFL-blowups at
the inflow/wall. But turbulunce in this viscous sublayer auto-adjusts
fast enough on itself, so we don't need to generate eddies.